+++ packages/app/src/styles/zIndex.js
@@ -1,0 +1,3 @@
+export const ZINDEX_SPINNER = 9000;
+export const ZINDEX_NOTIFICATION = 7000;
+export const ZINDEX_APP_MESSAGE = 6000;