+++ packages/firebase-hooks/src/firestore/useDoc.d.ts
@@ -1,10 +1,17 @@
-type Data = {
+interface Doc {
     id?: number;
     data?: Record<string, unknown>;
-};
+}
+
+type Data = Doc | Record<string, unknown>;
 
 export interface UseDocProps {
-    path?: string;
+    /** The firestore document path to fetch. */
+    path: string;
+    /** When set to true the document id will be merged with each document data as `id`. Otherwise each item in the array returned will be an object with id and data properties.  */
+    flat?: boolean;
+    /** When set to true it will not make the request. It is used for make conditional fetch. */
+    disabled?: boolean;
 }
 
 export default function (props: UseDocProps): [Data, boolean];