+++ packages/hooks/src/modal/useConnectModal.d.ts
@@ -1,6 +1,6 @@
 declare function useConnectModal(
-    modalId: any,
-    defaultProps: Record<string, unknown>,
+    modalId: string,
+    defaultProps?: Record<string, unknown>,
 ): Record<string, unknown>;
 
 export default useConnectModal;