+++ packages/app/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.29.0](https://github.com/nexxtway/rainbow-modules/compare/v0.28.1...v0.29.0) (2021-02-16)
+
+**Note:** Version bump only for package @rainbow-modules/app
+
+## [0.28.1](https://github.com/nexxtway/rainbow-modules/compare/v0.28.0...v0.28.1) (2021-02-15)
+
+### Bug Fixes
+
+-   add actions to d.ts file in app module ([37aef77](https://github.com/nexxtway/rainbow-modules/commit/37aef7768796098f02c04fecfaeb9535b466d82c))
+
+# [0.28.0](https://github.com/nexxtway/rainbow-modules/compare/v0.27.0...v0.28.0) (2021-02-15)
+
+### Bug Fixes
+
+-   update z-index on portals ([#328](https://github.com/nexxtway/rainbow-modules/issues/328)) ([dd73a64](https://github.com/nexxtway/rainbow-modules/commit/dd73a645fe7dd211541fbad23ae9d467ea57e549))
+
+### Features
+
+-   add isEditable prop to RecordField ([#326](https://github.com/nexxtway/rainbow-modules/issues/326)) ([c73e7da](https://github.com/nexxtway/rainbow-modules/commit/c73e7dae41412b60baa99604fa1faf7a053e9788)), closes [#320](https://github.com/nexxtway/rainbow-modules/issues/320)
+
+# [0.27.0](https://github.com/nexxtway/rainbow-modules/compare/v0.26.0...v0.27.0) (2021-02-08)
+
+### Bug Fixes
+
+-   forward ref passed through styled ([#322](https://github.com/nexxtway/rainbow-modules/issues/322)) ([640a7df](https://github.com/nexxtway/rainbow-modules/commit/640a7dfb7beba29b810a783efdcfcffe6725e760))
+
 # [0.26.0](https://github.com/nexxtway/rainbow-modules/compare/v0.25.0...v0.26.0) (2021-01-29)
 
 **Note:** Version bump only for package @rainbow-modules/app